+++ src/main/kotlin/com/dumbdogdiner/parkour/ParkourPlugin.kt
@@ -3,12 +3,10 @@ package com.dumbdogdiner.parkour
 import com.dumbdogdiner.parkour.commands.ParkourCommand
 import com.dumbdogdiner.parkour.courses.CourseManager
 import com.dumbdogdiner.parkour.editor.EditingSessionManager
-import com.dumbdogdiner.parkour.listeners.PlayerDropItemListener
-import com.dumbdogdiner.parkour.listeners.PlayerInteractListener
-import com.dumbdogdiner.parkour.listeners.PlayerQuitListener
-import com.dumbdogdiner.parkour.listeners.WorldListener
+import com.dumbdogdiner.parkour.listeners.*
 import com.dumbdogdiner.parkour.session.SessionManager
 import com.dumbdogdiner.parkour.utils.Configuration
+
 import org.bukkit.plugin.java.JavaPlugin
 
 class ParkourPlugin : JavaPlugin() {
@@ -28,8 +26,8 @@ class ParkourPlugin : JavaPlugin() {
         editingSessionManager = EditingSessionManager()
 
         // events events events
-        server.pluginManager.registerEvents(PlayerInteractListener(), this)
-        server.pluginManager.registerEvents(PlayerDropItemListener(), this)
+        server.pluginManager.registerEvents(PlayerSessionListener(), this)
+        server.pluginManager.registerEvents(PlayerMiscListener(), this)
         server.pluginManager.registerEvents(PlayerQuitListener(), this)
         server.pluginManager.registerEvents(WorldListener(), this)
